backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
qrcode[pil]
transformers
pypdf
pytesseract
Pillow
torch
python-multipart